Move sending message from InputBox component to InputPane
