+++ src/ui/components/input_box.rs
@@ -1,6 +1,6 @@
 use crossterm::event::{KeyCode, KeyEvent, KeyEventKind};
 use ratatui::{
-    prelude::{Backend, Rect},
+    prelude::Rect,
     style::{Color, Style, Stylize},
     widgets::{Block, Borders, Paragraph},
     Frame,
@@ -12,7 +12,7 @@ use crate::state::{action::Action, State};
 use super::{Component, ComponentRender};
 
 pub struct InputBox {
-    action_tx: UnboundedSender<Action>,
+    _action_tx: UnboundedSender<Action>,
     /// Current value of the input box
     text: String,
     /// Position of cursor in the editor area.
@@ -76,20 +76,15 @@ impl InputBox {
         }
     }
 
-    fn send_message(&mut self) {
-        let _ = self.action_tx.send(Action::SendMessage(self.text.clone()));
-        self.reset();
-    }
-
     fn clamp_cursor(&self, new_cursor_pos: usize) -> usize {
         new_cursor_pos.clamp(0, self.text.len())
     }
 }
 
 impl Component for InputBox {
-    fn new(_state: &State, action_tx: UnboundedSender<Action>) -> Self {
+    fn new(_state: &State, _action_tx: UnboundedSender<Action>) -> Self {
         Self {
-            action_tx: action_tx,
+            _action_tx,
             //
             text: String::new(),
             cursor_position: 0,
@@ -100,7 +95,7 @@ impl Component for InputBox {
         "Input"
     }
 
-    fn move_with_state(self, state: &State) -> Self
+    fn move_with_state(self, _state: &State) -> Self
     where
         Self: Sized,
     {
@@ -124,9 +119,6 @@ impl Component for InputBox {
             }
             KeyCode::Right => {
                 self.move_cursor_right();
-            }
-            KeyCode::Enter => {
-                self.send_message();
             }
             _ => {}
         }
